seguimos con los eliminar/listar

+++ app/Http/Controllers/LugarController.php
@@ -3,6 +3,8 @@
 namespace App\Http\Controllers;
 
 use App\Models\Lugar;
+use App\Models\Ruta;
+use Session;
 use Illuminate\Http\Request;
 use App\Http\Requests\StoreLugares;
 use App\Http\Requests\UpdateLugares;
@@ -27,7 +29,15 @@ class LugarController extends Controller
     }
 
     public function eliminarLugar(Lugar $lugar){
-      $lugar->delete();
+      $ruta = Ruta::where('origen_id', '=', $lugar->id, 'or', 'destino_id', '=', $lugar->id)->get()->first();
+      if (!empty($ruta)){
+        Session::flash('messageNO','El lugar se encuentra asignado a una ruta. Seleccione aquí para asignar un nuevo lugar.');
+        return redirect()->route('combi19.listarLugares')->with('ruta', $ruta)->with('lugar', $lugar);
+      }
+      else {
+        Session::flash('messageSI','El lugar se eliminó correctamente');
+        $lugar->delete();
+      }       
       return redirect()->route('combi19.listarLugares');
     }
